+++ gradio/_simple_templates/simpledropdown.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING, Any
 
 from gradio.components.base import Component, FormComponent
 from gradio.events import Events
+from gradio.i18n import I18nData
 
 if TYPE_CHECKING:
     from gradio.components import Timer
@@ -23,8 +24,8 @@ class SimpleDropdown(FormComponent):
         choices: list[str | int | float | tuple[str, str | int | float]] | None = None,
         *,
         value: str | int | float | Callable | None = None,
-        label: str | None = None,
-        info: str | None = None,
+        label: str | I18nData | None = None,
+        info: str | I18nData | None = None,
         every: Timer | float | None = None,
         inputs: Component | Sequence[Component] | set[Component] | None = None,
         show_label: bool | None = None,