+++ gradio/components/highlighted_text.py
@@ -10,6 +10,7 @@ from gradio_client.documentation import document
 from gradio.components.base import Component
 from gradio.data_classes import GradioModel, GradioRootModel
 from gradio.events import Events
+from gradio.i18n import I18nData
 
 if TYPE_CHECKING:
     from gradio.components import Timer
@@ -46,7 +47,7 @@ class HighlightedText(Component):
         show_inline_category: bool = True,
         combine_adjacent: bool = False,
         adjacent_separator: str = "",
-        label: str | None = None,
+        label: str | I18nData | None = None,
         every: Timer | float | None = None,
         inputs: Component | Sequence[Component] | set[Component] | None = None,
         show_label: bool | None = None,