+++ gradio/components/textbox.py
@@ -10,6 +10,7 @@ from gradio_client.documentation import document
 
 from gradio.components.base import Component, FormComponent
 from gradio.events import Events
+from gradio.i18n import I18nData
 
 if TYPE_CHECKING:
     from gradio.components import Timer
@@ -37,14 +38,14 @@ class Textbox(FormComponent):
 
     def __init__(
         self,
-        value: str | Callable | None = None,
+        value: str | I18nData | Callable | None = None,
         *,
         type: Literal["text", "password", "email"] = "text",
         lines: int = 1,
         max_lines: int | None = None,
-        placeholder: str | None = None,
-        label: str | None = None,
-        info: str | None = None,
+        placeholder: str | I18nData | None = None,
+        label: str | I18nData | None = None,
+        info: str | I18nData | None = None,
         every: Timer | float | None = None,
         inputs: Component | Sequence[Component] | set[Component] | None = None,
         show_label: bool | None = None,