+++ js/atoms/src/FullscreenButton.svelte
@@ -1,50 +1,25 @@
 <script lang="ts">
-	import { onMount, createEventDispatcher } from "svelte";
+	import { createEventDispatcher } from "svelte";
 	import { IconButton } from "@gradio/atoms";
 	import { Maximize, Minimize } from "@gradio/icons";
 
-	export let container: HTMLElement | undefined = undefined;
 	const dispatch = createEventDispatcher<{
-		fullscreenchange: boolean;
+		fullscreen: boolean;
 	}>();
 
-	export let is_full_screen = false;
-
-	onMount(() => {
-		document.addEventListener("fullscreenchange", () => {
-			is_full_screen = !!document.fullscreenElement;
-			dispatch("fullscreenchange", is_full_screen);
-		});
-	});
-
-	const toggle_full_screen = async (): Promise<void> => {
-		if (!container) return;
-
-		if (!is_full_screen) {
-			try {
-				await container.requestFullscreen();
-			} catch (e) {
-				console.error(e);
-			}
-		} else {
-			await document.exitFullscreen();
-			is_full_screen = !is_full_screen;
-		}
-	};
+	export let fullscreen;
 </script>
 
-{#if !is_full_screen}
-	<IconButton
-		Icon={Maximize}
-		label="View in full screen"
-		on:click={toggle_full_screen}
-	/>
-{/if}
-
-{#if is_full_screen}
+{#if fullscreen}
 	<IconButton
 		Icon={Minimize}
-		label="Exit full screen"
-		on:click={toggle_full_screen}
+		label="Exit fullscreen mode"
+		on:click={() => dispatch("fullscreen", false)}
+	/>
+{:else}
+	<IconButton
+		Icon={Maximize}
+		label="Fullscreen"
+		on:click={() => dispatch("fullscreen", true)}
 	/>
 {/if}