+++ js/core/src/Blocks.svelte
@@ -29,8 +29,6 @@
 	} from "@gradio/client";
 	import * as screen_recorder from "./screen_recorder";
 
-	setupi18n();
-
 	export let root: string;
 	export let components: ComponentMeta[];
 	export let layout: LayoutNode;
@@ -54,6 +52,8 @@
 	export let max_file_size: number | undefined = undefined;
 	export let initial_layout: ComponentMeta | undefined = undefined;
 	export let css: string | null | undefined = null;
+
+	setupi18n(app.config?.i18n_translations ?? undefined);
 
 	let {
 		layout: _layout,