+++ js/dropdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/dropdown
 
+## 0.9.21
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.11
+- @gradio/atoms@0.16.1
+
 ## 0.9.20
 
 ### Dependency updates