+++ js/simpledropdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/simpledropdown
 
+## 0.3.21
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.11
+- @gradio/atoms@0.16.1
+
 ## 0.3.20
 
 ### Dependency updates