+++ js/simpleimage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/simpleimage
 
+## 0.8.29
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.11
+- @gradio/atoms@0.16.1
+- @gradio/client@1.15.0
+- @gradio/upload@0.16.5
+
 ## 0.8.28
 
 ### Dependency updates